Big tidyup, simpler now
